faceRecognitionAPI/requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
canvasapi==3.0.0
cement==2.8.2
certifi==2022.9.14
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cmake==3.24.1
colorama==0.4.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.1.1
django-cors-headers==3.10.0
django-environ==0.8.1
django-eventstream==4.4.1
django-filter==22.1
django-grip==3.2.0
django-storages==1.13.1
djangorestframework==3.14.0
dlib==19.24.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.16.0
gripcontrol==4.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
incremental==22.10.0
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.4
msgpack-python==0.5.6
mysqlclient==2.1.0
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.2.0
pubcontrol==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==5.4.1
redis==4.3.4
requests==2.26.0
semantic-version==2.8.5
service-identity==21.1.0
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.5.1
